+++ services/product/src/main/java/com/teamz/product/product/ProductRequest.java
@@ -12,7 +12,7 @@ public record ProductRequest(
         @NotNull(message = "Product description is required")
         String description,
         @Positive(message = "Available quantity should be positive")
-        double availableQuantity,
+        Long availableQuantity,
         @Positive(message = "Price should be positive")
         BigDecimal price,
         @NotNull(message = "Product category is required")
